Rename and move Cardano.CLI.EraBased.Commands.Address to Cardano.CLI.Commands.Address
